add normal command fix Setting.h

+++ arduino_climat_controll_program/Setting.h
@@ -108,7 +108,7 @@ class Setting {
 			printMode();
 		}
 		
-		void setWantTempInCar(int temp) {
+		void setWantTempInCar(float temp) {
 			data.wantTempInCar = temp;
 			printWantTempInCar();
 		}
